test(quality): 分区动量用例移出三渲二文件
这两条只用 PIL 和 slicing.quality,与三渲二 provider 无关。留在
test_render3d_route_and_assets.py 里会被那个文件的整体 skip 一并带走,
于是本 PR 新增的这项读数在 provider 合入前一条断言都不跑。
